requirements.txt
ffmpeg-python
PyPDF2
python-docx
Pillow
